refactor(core): split executor helpers into one declaration per file
