+++ jindong-core/src/commonMain/kotlin/io/github/compose/jindong/core/executor/HandleExpiry.kt
@@ -13,8 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@file:Suppress("ktlint:standard:filename")
-
 package io.github.compose.jindong.core.executor
 
 import kotlin.time.Duration
